Inventaire : UID objet en long + strip marqueur OAK + format OQ correct
Sur Hystoria les UID d'instance objet dépassent Int32 (ex. 0x1dc7e9d08
= 7 994 252 552). L'ancien int.TryParse en hex overflowait → id=0 →
objet jeté → 0 objets dans l'inventaire après login OU après récolte.
Le préfixe « O » du marqueur d'entrée OAK (ex. « O1dc7e9d08~12f~1~~ »)
n'était pas stripé non plus → ParseHex=0 → même jet.

Fix :
- ObjetInventaire.Identifiant + ObjetParse.Identifiant : int → long.
- ParseHex / ParseHexLong renvoient long, strip 1er char si non-hex.
- MessageObjetRetrait/Quantite : id long, lecture après dernier '|'.
- MessageObjetQuantite : préfixe « Oq » (jamais reçu) → « OQ » réel +
  format « <charGID>|<uid>,<qty> » (capture loot Hystoria).
- ApiAnka.objectUid : retour long (l'expression `?.Identifiant ?? -1`
  est désormais long).

Validé prod (log 06:47:19.840 : « [INV] Inventaire initial chargé : 84
objets », OQ401770|7994252552,4 reçu et appliqué). Avant : toujours 0.

+++ Divers/Jeu/Personnage/Personnage.cs
@@ -116,7 +116,9 @@ public sealed class Personnage
 
 public sealed class ObjetInventaire
 {
-    public int Identifiant { get; set; }
+    // UID d'instance objet Hystoria : dépasse Int32 (ex. 0x1dc7e9d08 =
+    // 7 994 252 552). DOIT être long sinon overflow → objet jamais ajouté.
+    public long Identifiant { get; set; }
     public int IdTemplate { get; set; }
     public int Quantite { get; set; } = 1;
     public int Position { get; set; }
